requirements.txt
boto3==1.15.11
botocore==1.18.11
certifi==2020.6.20
chardet==3.0.4
idna==2.10
jmespath==0.10.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
urllib3==1.25.10
